+++ pysat/_instrument.py
@@ -1106,8 +1106,14 @@ class Instrument(object):
                     self.data[key] = in_data
                 elif len(np.shape(in_data)) <= 1:
                     # If not an xarray input, but still iterable, then we
-                    # go through to process the 1D input
-                    if np.shape(in_data) == np.shape(self.index):
+                    # go through to process the input
+                    if key in self.variables and (
+                            np.shape(in_data) == np.shape(self.data[key])):
+                        # The ND input has the same shape as the current data
+                        # and can be assigned directly without adjusting the
+                        # dimensions. Only works with existing data.
+                        self.data[key] = (self.data[key].dims, in_data)
+                    elif np.shape(in_data) == np.shape(self.index):
                         # 1D input has the correct length for storage along
                         # 'Epoch'.
                         self.data[key] = (epoch_name, in_data)